Significant rewrite now allows non-ordered search terms
Original version only supported ordered search, that is,
windows always started with the first term specified.
Now it is able to find windows with any term first.
However, after playing with it, I found that ordered
search was usually the way I expected and preferred
it to work.  So I made --ordered the default.
--no-ordered is available if order doesn't matter.

The re-write made the code simpler and cleaner anyway.